Release/candidate2.13

See merge request !1801

+++ kadena2.0/Kadena2.0.Dto/CustomerData/UserBudgetDto.cs
@@ -5,6 +5,6 @@ namespace Kadena.Dto.CustomerData
     public class UserBudgetDto
     {
         public int ItemID { get; set; }
-        public int UserBudget { get; set; }
+        public decimal UserBudget { get; set; }
     }
 }
